barbican: split existing command in a single file each
preparatory work for more complex command set (e.g. sdk)
